requirements.txt
asgiref==3.11.0
dj-database-url==3.0.1
Django==5.0.14
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.4
tzdata==2025.2
